Move the FastTimerService code base into the plugins directory
