--- HLTrigger/Timer/plugins/plugins.cc
@@ -1,5 +1,0 @@
-#include "FWCore/ServiceRegistry/interface/ServiceMaker.h"
-
-// declare FastTimerService as a framework Service
-#include "HLTrigger/Timer/interface/FastTimerService.h"
-DEFINE_FWK_SERVICE(FastTimerService);
